requirements.txt
fastapi
httpie
strawberry-graphql
uvicorn
